chore(ourlogs): Separate explore page from logs page
