+++ static/app/views/explore/utils.tsx
@@ -1,9 +1,10 @@
 import type {Location} from 'history';
 import * as qs from 'query-string';
 
+import {t} from 'sentry/locale';
 import type {PageFilters} from 'sentry/types/core';
 import type {Series} from 'sentry/types/echarts';
-import type {Confidence} from 'sentry/types/organization';
+import type {Confidence, Organization} from 'sentry/types/organization';
 import type {Project} from 'sentry/types/project';
 import {defined} from 'sentry/utils';
 import {DiscoverDatasets} from 'sentry/utils/discover/types';
@@ -115,4 +116,47 @@ export function viewSamplesTarget(
     mode: Mode.SAMPLES,
     query: search.formatString(),
   });
+}
+
+export type MaxPickableDays = 7 | 14 | 30;
+export type DefaultPeriod = '7d' | '14d' | '30d';
+
+export function limitMaxPickableDays(organization: Organization): {
+  defaultPeriod: DefaultPeriod;
+  maxPickableDays: MaxPickableDays;
+  relativeOptions: Record<string, React.ReactNode>;
+} {
+  const defaultPeriods: Record<MaxPickableDays, DefaultPeriod> = {
+    7: '7d',
+    14: '14d',
+    30: '30d',
+  };
+
+  const relativeOptions: [DefaultPeriod, React.ReactNode][] = [
+    ['7d', t('Last 7 days')],
+    ['14d', t('Last 14 days')],
+    ['30d', t('Last 30 days')],
+  ];
+
+  const maxPickableDays: MaxPickableDays = organization.features.includes(
+    'visibility-explore-range-high'
+  )
+    ? 30
+    : organization.features.includes('visibility-explore-range-medium')
+      ? 14
+      : 7;
+  const defaultPeriod: DefaultPeriod = defaultPeriods[maxPickableDays];
+
+  const index = relativeOptions.findIndex(([period, _]) => period === defaultPeriod) + 1;
+  const enabledOptions = relativeOptions.slice(0, index);
+
+  return {
+    defaultPeriod,
+    maxPickableDays,
+    relativeOptions: {
+      '1h': t('Last hour'),
+      '24h': t('Last 24 hours'),
+      ...Object.fromEntries(enabledOptions),
+    },
+  };
 }
